trino-query/requirements.txt
trino>=0.330.0
python-dotenv>=1.0.0
tabulate>=0.9.0
pytest>=8.0.0  # For automated validation
